* Add some rate-limiting in sample backend scripts * Configure the pid file path via the config file now * Make the slabdb debug logs supported by the debug flags * Start going over all uses of ino_t and off_t, and other types, to guard   against integer overflow bugs. Our max inodes and file offsets are now   off_t (that is, ULONG_MAX aka 2^63). * Added a "fix" flag to fsck to start tackling the issue of invalid inodes   inside dir_entry's * A bit more logging, a bit of error handling; we'll have to eventually   send the manager's errors back to the client for clarity * Added the "shutdown" and "inode_tables" commands to potatoctl * More details in potatoctl status * Fixed a few memory leaks using valgrind * I think we're at v0.2 now.

+++ slabdb.h
@@ -53,9 +53,9 @@ int slabdb_get_next_itbl(off_t *, struct exlog_err *);
 int slabdb_loop(int(*)(const struct slab_key *, const struct slabdb_val *,
         void *), void *, struct exlog_err *);
 
-int      slabdb_begin_txn(struct exlog_err *);
-int      slabdb_commit_txn(struct exlog_err *);
-int      slabdb_rollback_txn(struct exlog_err *);
-uint64_t slabdb_count(struct exlog_err *);
+int     slabdb_begin_txn(struct exlog_err *);
+int     slabdb_commit_txn(struct exlog_err *);
+int     slabdb_rollback_txn(struct exlog_err *);
+ssize_t slabdb_count(struct exlog_err *);
 
 #endif
